Remember: It is impossible to make anything foolproof because fools are so ingenious.
This code had a string meant for translators used like i18n("Translator:
please read xxx file");
And (at least) the spanish translator simply translated that message instead
of reading that file, leaving some translations completely broken.  Grrr.
Let's see if they note that the message is for them this way.

Btw, why would anyone want to convert the string that was only meant to translators
to local8Bit and then to a STL string, if the string is not used at all?

svn path=/trunk/kdeadmin/kcron/; revision=217305