Port module to QML
Porting kcron to QML. Tables have been separated into a new plugin
from Kirigami Addons. The module's interaction with crontab (crontablib)
has not changed.

Changed behavior when there are no cron jobs: previously, a dialog box
appeared to welcome the user, now all information is displayed on a
separate tab of the module.

The main page of the module has remained almost unchanged. The "Print"
button has been moved to the ToolBar because not only tasks, but also
variables were saved in the PDF, and leaving it next to the task table,
is not correct, since it can mislead the user as to what exactly will
be saved. The appearance of the "Status" column has also been simplified.

Previously, changing/creating variables or tasks was implemented through
modal windows; now separate pages are used.

The old implementation of printing to PDF used QPainter and all rendering
was based on QWidgets. Because of this it was decided to print HTML code,
which also corrected some printing artifacts.